backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
alembic==1.0.5
apispec-marshmallow-advanced==0.4
apispec==1.1.0            # via apispec-marshmallow-advanced, hapic
attrs==18.2.0             # via cmd2
babel==2.6.0
beaker==1.10.0
beautifulsoup4==4.6.3
caldav==0.6.0
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via rq
cliff==2.14.0
cmd2==0.9.6               # via cliff
colorama==0.4.1           # via cmd2
colour==0.1.5
defusedxml==0.5.0         # via wsgidav
elasticsearch-dsl==7.0.0
elasticsearch==7.0.1
email-reply-parser==0.5.9
filedepot==0.5.2
filelock==3.0.10
hapic[marshmallow]==0.75
hupper==1.4.2             # via pyramid
idna==2.8                 # via requests
imapclient==2.1.0
jinja2==2.10              # via pyramid-jinja2
jsmin==2.2.2              # via wsgidav
ldap3==2.5.1              # via pyramid-ldap3
lxml==4.2.5
mako==1.0.7
markdown==3.0.1
markupsafe==1.1.0         # via jinja2, mako, pyramid-jinja2
marshmallow==2.16.3
multidict==4.5.2          # via hapic
packaging==18.0           # via preview-generator
pastedeploy==2.0.0        # via plaster-pastedeploy
pathlib==1.0.1            # via preview-generator
pbr==5.1.1                # via cliff, stevedore
pdf2image==1.1.0          # via preview-generator
pillow==5.3.0             # via pdf2image, preview-generator
plaster-pastedeploy==0.6
plaster==1.0              # via plaster-pastedeploy, pyramid
prettytable==0.7.2        # via cliff
preview-generator==0.10
pyasn1==0.4.4             # via ldap3
pyexifinfo==0.4.0         # via preview-generator
pygments==2.3.0           # via pyramid-debugtoolbar
pyparsing==2.3.0          # via cliff, packaging
pypdf2==1.26.0            # via preview-generator
pyperclip==1.7.0          # via cmd2
pyramid-beaker==0.8
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.7
pyramid-ldap3==0.4
pyramid-mako==1.0.2
pyramid-multiauth==0.9.0
pyramid-retry==1.0
pyramid-tm==2.2.1
pyramid==1.10.1
python-dateutil==2.7.5
python-editor==1.0.3      # via alembic
python-magic==0.4.15      # via preview-generator
python-slugify==1.2.6
pytz==2018.7              # via babel, caldav, tzlocal
pyyaml==3.13
radicale==2.1.11
redis==2.10.6
repoze.lru==0.7           # via pyramid-debugtoolbar
requests==2.21.0
rq==0.12.0
six==1.11.0               # via caldav, cliff, elasticsearch-dsl, imapclient, packaging, python-dateutil, stevedore
sqlalchemy==1.2.14
stevedore==1.30.0         # via cliff
transaction==2.4.0
translationstring==1.3    # via pyramid
tzlocal==1.5.1            # via caldav
unidecode==1.0.23         # via filedepot, python-slugify
urllib3==1.24.1           # via elasticsearch, requests
venusian==1.1.0           # via pyramid
vobject==0.9.6.1          # via caldav, radicale
waitress==1.1.0
wand==0.4.5               # via preview-generator
wcwidth==0.1.7            # via cmd2
webob==1.8.4              # via pyramid
wsgidav==2.4.1
xvfbwrapper==0.2.9        # via preview-generator
zope.deprecation==4.3.0   # via pyramid, pyramid-jinja2
zope.interface==4.6.0     # via pyramid, pyramid-retry, transaction, zope.sqlalchemy
zope.sqlalchemy==1.0
